+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.19.0] - 2021-12-13
 ### Added
 - Edit deck button.
 - Edit model button.
@@ -464,7 +466,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Hydration logic on production environment
 
 
-[Unreleased]: https://github.com/cramkle/cramkle/compare/v0.18.0...HEAD
+[Unreleased]: https://github.com/cramkle/cramkle/compare/v0.19.0...HEAD
+[0.19.0]: https://github.com/cramkle/cramkle/compare/v0.18.0...v0.19.0
 [0.18.0]: https://github.com/cramkle/cramkle/compare/v0.17.0...v0.18.0
 [0.17.0]: https://github.com/cramkle/cramkle/compare/v0.16.0...v0.17.0
 [0.16.0]: https://github.com/cramkle/cramkle/compare/v0.15.2...v0.16.0